requirements.txt
blinker==1.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
gspread==5.12.4
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
oauthlib==3.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.25.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
urllib3==2.1.0
wadllib==1.3.6
zipp==1.0.0
